Version 1.0
This is the original code of the game from when I created it in the summer of 2017. At that time, I did not know how to use Github, thus the 1.5 years between creation and publication.

This version is fully functional.

hexagonal_2048.js is unused.

The primary goal of subsequent commits is to rewrite the project to be consistent with what I learn about good practice in object-oriented programming.